requirements.txt
Flask==3.0.3
tensorflow-cpu==2.12.0
numpy==1.23.5
Pillow==9.5.0
gunicorn==21.2.0
